+++ src/main/java/com/polar_moviechart/movieservice/domain/service/dtos/StatDto.java
@@ -9,18 +9,18 @@ import java.time.LocalDate;
 @Getter
 public class StatDto {
     private LocalDate date;
-    private Integer value;
+    private long value;
 
-    public StatDto(LocalDate date, Integer value) {
+    public StatDto(LocalDate date, Long value) {
         this.date = date;
         this.value = value;
     }
 
     public static StatDto from(MovieDailyStat stat, StatType statType) {
         if (statType == StatType.AUDIENCE) {
-            return new StatDto(stat.getDate(), stat.getAudience());
+            return new StatDto(stat.getDate(), stat.getAudience().longValue());
         } else if (statType == StatType.RANKING) {
-            return new StatDto(stat.getDate(), stat.getRanking());
+            return new StatDto(stat.getDate(), stat.getRanking().longValue());
         } else {
             return new StatDto(stat.getDate(), stat.getRevenue());
         }
